8 dec

+++ src/main/java/com/ServiceMyVehicleProject/demo/Service/MechanicShopService.java
@@ -14,6 +14,9 @@ public class MechanicShopService {
 	@Autowired
 	private MechanicShopRepo shopdao;
 	
+	@Autowired
+	private ShopAddressService shopaddser;
+	
 	public List<MechanicShop> getAllShops(){
 		return shopdao.findAll();
 	}
@@ -23,15 +26,23 @@ public class MechanicShopService {
 	}
 	
 	public MechanicShop createShop(MechanicShop shop) {
+		shopaddser.createShopAddress(shop.getShopaddress());
 		return shopdao.save(shop);
 	}
 	
 	public MechanicShop updateShop(MechanicShop shop) {
+		shopaddser.updateShopAddress(shop.getShopaddress());
 		return shopdao.save(shop);
 	}
 	
 	public Boolean deleteShopById(long id) {
-		shopdao.deleteById(id);
+		MechanicShop shop = shopdao.findById(id).orElse(null);
+		if(shop==null) {
+			return true;
+		}else {
+			shopdao.deleteById(id);
+			shopaddser.deleteShopAddress(shop.getShopaddress().getSaid());
+		}
 		return true;
 	}
 }
